Added high-dpi / retina support to the no-game screen; additional minor tweaks and cleanup.

+++ src/modules/graphics/Quad.cpp
@@ -29,7 +29,7 @@ namespace love
 namespace graphics
 {
 
-Quad::Quad(const Quad::Viewport &v, float sw, float sh)
+Quad::Quad(const Quad::Viewport &v, double sw, double sh)
 	: sw(sw)
 	, sh(sh)
 {
@@ -41,7 +41,7 @@ Quad::~Quad()
 {
 }
 
-void Quad::refresh(const Quad::Viewport &v, float sw, float sh)
+void Quad::refresh(const Quad::Viewport &v, double sw, double sh)
 {
 	viewport = v;
 
@@ -53,20 +53,20 @@ void Quad::refresh(const Quad::Viewport &v, float sw, float sh)
 	vertices[0].x = 0.0f;
 	vertices[0].y = 0.0f;
 	vertices[1].x = 0.0f;
-	vertices[1].y = v.h;
-	vertices[2].x = v.w;
+	vertices[1].y = (float) v.h;
+	vertices[2].x = (float) v.w;
 	vertices[2].y = 0.0f;
-	vertices[3].x = v.w;
-	vertices[3].y = v.h;
+	vertices[3].x = (float) v.w;
+	vertices[3].y = (float) v.h;
 
-	vertices[0].s = v.x/sw;
-	vertices[0].t = v.y/sh;
-	vertices[1].s = v.x/sw;
-	vertices[1].t = (v.y+v.h)/sh;
-	vertices[2].s = (v.x+v.w)/sw;
-	vertices[2].t = v.y/sh;
-	vertices[3].s = (v.x+v.w)/sw;
-	vertices[3].t = (v.y+v.h)/sh;
+	vertices[0].s = (float) (v.x/sw);
+	vertices[0].t = (float) (v.y/sh);
+	vertices[1].s = (float) (v.x/sw);
+	vertices[1].t = (float) ((v.y+v.h)/sh);
+	vertices[2].s = (float) ((v.x+v.w)/sw);
+	vertices[2].t = (float) (v.y/sh);
+	vertices[3].s = (float) ((v.x+v.w)/sw);
+	vertices[3].t = (float) ((v.y+v.h)/sh);
 }
 
 void Quad::setViewport(const Quad::Viewport &v)
